Add https:// to front of domain argument if it's not a URL
The "domain" argument to the 2FAS API isn't actually expecting a
domain, it's expecting a URL (see
https://github.com/twofas/2fas-server/issues/6).

Arguably it would be better to rename the "--domain" argument in this
script to "--url" for clarity, but absent that, at the very least we
should check if the user specifies a domain instead of a URL—since
that's what the parameter is called, it's an easy mistake to make!—and
if so, then add "https://" to the beginning of it so that we're
passing the 2FAS server something it expects.

Fixes https://github.com/dspitzer/twofas-cli/issues/1.